написал тесты для каталога и имплементации дао. Написал имплементации дао для инвентору

+++ catalog/src/main/java/com/netcracker/zagursky/dao/OfferDao.java
@@ -1,6 +1,5 @@
 package com.netcracker.zagursky.dao;
 
-import com.netcracker.zagursky.entity.Category;
 import com.netcracker.zagursky.entity.Offer;
 
 import java.util.List;
@@ -8,8 +7,10 @@ import java.util.List;
 /**
  * Created by Dzenyaa on 08.11.2017.
  */
-public interface OfferDao extends GenericDao<Offer,Long> {
-     Offer findByName(String name);
-     List<Offer> findByTag(String tagName);
-     List<Offer> findByCategory(String categoryName);
+public interface OfferDao extends GenericDao<Offer, Integer> {
+    Offer findByName(String name);
+
+    List<Offer> findByTag(String tagName);
+
+    List<Offer> findByCategory(String categoryName);
 }
